Set the default data directory correctly
Previously, we set the default to the empty string "" and manually set
the default directory in `node.New`. This meant the incorrect default
was printed in two cases:

1. `--help`
2. When printing the config on boot.

Furthermore, this centralizes all of our config management to the
`cmd/juno` package; if `juno.NewCmd` returns without an error, the
config is structurally valid.